% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupCDM.r
\name{backupCDM}
\alias{backupCDM}
\title{Backup Vocabulary and Event tables before pruning.}
\usage{
backupCDM(connectionDetails,cdmDatabaseSchema,vocabDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the CDM
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabDatabaseSchema}{The name of the database schema that contains the Vocabulary
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'vocab_instance.dbo'.}
}
\description{
This function performs a "back up" of the original tables by renaming them.
}
