% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCDMIndexAndConstraintScripts.r
\name{CreateCDMIndexAndConstraintScripts}
\alias{CreateCDMIndexAndConstraintScripts}
\title{Create Index and Constraint DDL SQL scripts for CDM tables on rdbms that support these features.}
\usage{
CreateCDMIndexAndConstraintScripts(connectionDetails,cdmSchema,cdmVersion)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the CDM database schema.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{Your CDM version.  Currently "5.3.1" and "6.0.0" are supported.}
}
\description{
This function creates one or more SQL scripts as defined in https://github.com/OHDSI/CommonDataModel.
}
\details{
This function creates SQL scripts for the indices and constraints on tables in a CDM by referring to the
         correct SQL DDL script in the OHDSI CommonDataModel repo. The database platform is 
         determined by \code{connectionDetails$dbms}.  Currently "oracle", "postgresql", "pdw", and "netezza" are supported.
         The SQL DDL scripts are written to the \code{output} directory.  The SQL scripts can then be run manually in a SQL session
         or by using \code{DatabaseConnector::executeSql(connection,DDLscriptName)}
}
