% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtraIndices.R
\name{createExtraIndices}
\alias{createExtraIndices}
\title{Create Optional Extra Indices for ETL Performance}
\usage{
createExtraIndices(
  connectionDetails,
  cdmSchema,
  syntheaSchema,
  syntheaVersion,
  outputFolder = NULL,
  sqlOnly = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the CDM database schema. Requires read and write permissions to this schema. On SQL
Server, this should specify both the database and the schema, so for example 'cdm_instance.dbo'.}

\item{syntheaSchema}{The name of the Synthea database schema. Requires read and write permissions to this schema. On SQL
Server, this should specify both the database and the schema, so for example 'synthea.dbo'.}

\item{syntheaVersion}{Your Synthea version. Currently "2.7.0" and "3.0.0" are supported.}

\item{outputFolder}{Location of the SQL scripts if sqlOnly = TRUE. Default is NULL.}

\item{sqlOnly}{A boolean that determines whether to create the indices or generate a SQL scripts. Default is FALSE.}
}
\description{
This function creates indices for certain tables which may help to speed up LoadEventTables.
}
\details{
This function creates indices which have been found to speed up certain long-running INSERT queries in LoadEventTables,
         for some users. Indices are created on the intermediate vocabulary mapping tables; the person & provider CDM tables;
         and the claims_transactions Synthea table (in Synthea 3.0.0).
}
