% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropMapAndRollupTables.R
\name{DropMapAndRollupTables}
\alias{DropMapAndRollupTables}
\title{Drop Vocabulary Mapping And Visit Rollup Tables.}
\usage{
DropMapAndRollupTables(connectionDetails, cdmSchema)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that will contain the Vocab mapping
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}
}
\description{
This function cleans up the tables created by the \cr\code{CreateVocabMapTables()} and \cr\code{CreateVisitRollupTables()}
}
\details{
This function assumes \cr\code{CreateVocabMapTables()} and \cr\code{CreateVisitRollupTables()} have already been run.
}
