% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETLSyntheaBuilder.R
\docType{package}
\name{ETLSyntheaBuilder-package}
\alias{ETLSyntheaBuilder}
\alias{ETLSyntheaBuilder-package}
\title{ETLSyntheaBuilder: A Builder for Converting the Synthea Data to the OMOP CDM}
\description{
ETL and Builder to convert Synthea Data to the OMOP CDM. Staring with csv files for an OMOP Vocabulary and csv files for Synthea, this package creates database tables from these csv files and maps them correctly to create a full functional OMOP CDM from Synthea. As the CDM Vocabulary can be quite large, there is the option to copy the vocabulary from another schema that already has it loaded. This package assumes an RDBMS exists to which you have read/write access.
}
\keyword{internal}
