% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventConceptId.r
\name{getEventConceptId}
\alias{getEventConceptId}
\title{Get All Concept_ids In All Event Tables.}
\usage{
getEventConceptId(connectionDetails,cdmDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the CDM
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}
}
\description{
This function gets the unique concept_ids in all event tables.
}
