% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCDMTables.r
\name{LoadCDMTables}
\alias{LoadCDMTables}
\title{Load CDM Tables.}
\usage{
LoadCDMTables (connectionDetails, cdmDatabaseSchema, syntheaDatabaseSchema)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that will contain the different VISIT
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{syntheaDatabaseSchema}{The name of the database schema that contain the Synthea
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabDatabaseSchema}{The name of the database schema that contains the OMOP Vocabulary tables. Requires
read permissions to this database. By default this is set to cdmDatabaseSchema.}
}
\description{
This function loads the CDM tables with Synthea data.  The tables that comprise the Vocabulary are
              loaded via \cr\code{LoadVocabTables()}.
}
\details{
This function assumes \cr\code{createCDMTables()}, \cr\code{createSyntheaTables()}, \cr\code{LoadSyntheaTables()},
             and \cr\code{LoadVocabTables()} have all been run.
}
