% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSyntheaTables.r
\name{LoadSyntheaTables}
\alias{LoadSyntheaTables}
\title{Load Synthea Tables.}
\usage{
LoadSyntheaTables(connectionDetails, syntheaSchema, syntheaFileLoc)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{syntheaSchema}{The name of the database schema that will contain the Synthea
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{syntheaFileLoc}{The location of the Synthea csv files created by running the executable run_synthea.}
}
\description{
This function populates all Synthea tables.
}
\details{
This function assumes \cr\code{createSyntheaTables()} has already been run.  Additionally,
             this function assumes the data files used to populate the tables are csv files generated by 
             running the basic setup (creating a database with 1000 people):    
                 \cr\code{git clone https://github.com/synthetichealth/synthea.git}
                 \cr\code{cd synthea}
                 \cr\code{./gradlew build check test}          
                 \cr\code{./run_synthea -p 1000}

              You can enable csv records in src/main/resources/synthea.properties by setting exporter.csv.export = true.
              As of the time of this writing the csv files can be found at synthe/output/csv.
              For more details: \href{https://github.com/synthetichealth/synthea/wiki/Basic-Setup-and-Running}{Synthea Basic Setup}
}
