% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadVocabFromSchema.r
\name{LoadVocabFromSchema}
\alias{LoadVocabFromSchema}
\title{Load Vocabulary Tables From a Preexisting Schema.}
\usage{
LoadVocabFromSchema(connectionDetails, cdmSourceSchema, cdmTargetSchema)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSourceSchema}{The name of the database schema that already contains the Vocabulary
tables to copy.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmTargetSchema}{The name of the database schema into which to copy the Vocabulary
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}
}
\description{
This function populates all Vocabulary tables with data from Vocabulary tables in a specified schema.
}
\details{
This function assumes \cr\code{createCDMTables()} has already been run and \cr\code{cdmSourceSchema} has all required Vocabulary tables.
}
