% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadVocabTables.r
\name{LoadVocabTables}
\alias{LoadVocabTables}
\title{Load Vocabulary Tables.}
\usage{
LoadVocabTables(connectionDetails, cdmDatabaseSchema, vocabFileLoc)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that will contain the Vocabulary 
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabFileLoc}{The location of the vocabulary csv files.}
}
\description{
This function populates all Vocabulary tables.
}
\details{
This function assumes \cr\code{createCDMTables()} has already been run.
}
