% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneCDM.r
\name{pruneCDM}
\alias{pruneCDM}
\title{Prune a CDM.}
\usage{
pruneCDM(connectionDetails,cdmDatabaseSchema,vocabDatabaseSchema,pathToSql = NULL)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the CDM
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabDatabaseSchema}{The name of the database schema that contains the Vocabulary
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'vocab_instance.dbo'.}

\item{pathToSql}{Optional location of the sql files (defaults to inst/sql/sql_server)}
}
\description{
This function is simply a convenience wrapper for the other calls.
}
