% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncateVocabTables.r
\name{TruncateVocabTables}
\alias{TruncateVocabTables}
\title{Truncate Vocabulary Tables.}
\usage{
TruncateVocabTables(connectionDetails, cdmSchema)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that contains the Vocabulary (and CDM).
Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}
}
\description{
This function truncates the Vocabulary tables.
}
