% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCDMTables.r
\name{CreateCDMTables}
\alias{CreateCDMTables}
\title{Create Common Data Model Tables.}
\usage{
CreateCDMTables(
  connectionDetails,
  cdmSchema,
  cdmVersion,
  outputFolder = NULL,
  sqlOnly = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the CDM database schema.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{Your CDM version.  Currently "5.3" and "5.4" are supported.}

\item{outputFolder}{Location of the SQL scripts if sqlOnly = TRUE. Default is NULL.}

\item{sqlOnly}{A boolean that determines whether or not to perform the load or generate SQL scripts. Default is FALSE.}
}
\description{
This function creates the all the required tables for the CDM.
}
\details{
This function creates all the tables in a CDM by calling \code{CommonDataModel::executeDdl()}.
         Supported CDM versions and db dialects are determined by \code{CommonDataModel::listSupportedVersions()}
         and \code{CommonDataModel::listSupportedDialects()}, respectively.
}
