% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCDMTables.r
\name{CreateCDMTables}
\alias{CreateCDMTables}
\title{Create Common Data Model Tables.}
\usage{
CreateCDMTables(connectionDetails,cdmSchema,cdmVersion)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the CDM database schema.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{Your CDM version.  Currently "5.3.1" and "6.0.0" are supported.}
}
\description{
This function creates the all the required tables for the CDM.
}
\details{
This function creates all the tables in a CDM by referring to the
         correct SQL DDL script in the OHDSI CommonDataModel repo. The database platform is 
         determined by \code{connectionDetails$dbms}.  Currently "oracle", "postgresql", "pdw", 
         "impala", "netezza", "bigquery", and "redshift" are supported.
}
