% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSyntheaTables.r
\name{CreateSyntheaTables}
\alias{CreateSyntheaTables}
\title{Create Synthea Tables.}
\usage{
CreateSyntheaTables(connectionDetails, syntheaSchema, syntheaVersion = "2.7.0")
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{syntheaSchema}{The name of the database schema that will contain the Synthea
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{syntheaVersion}{The version of Synthea used to generate the csv files.
Currently "2.7.0" only is supported.}
}
\description{
This function creates all Synthea tables.
}
