% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVisitRollupTables.r
\name{CreateVisitRollupTables}
\alias{CreateVisitRollupTables}
\title{Create Tables for Visit Rollup Logic.}
\usage{
CreateVisitRollupTables(
  connectionDetails,
  cdmSchema,
  syntheaSchema,
  cdmVersion,
  sqlOnly = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that will contain the different VISIT
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{syntheaSchema}{The name of the database schema that contain the Synthea
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'synthea_instance.dbo'.}

\item{cdmVersion}{The version of your CDM.  Currently "5.3" and "5.4" are supported.}

\item{sqlOnly}{A boolean that determines whether or not to perform the load or generate SQL scripts. Default is FALSE.}
}
\description{
This function creates tables ALL_VISITS, ASSIGN_ALL_VISIT_IDS, and FINAL_VISIT_IDS.
}
\details{
This function assumes \cr\code{createCDMTables()}, \cr\code{createSyntheaTables()}, \cr\code{LoadSyntheaTables()},
         have all been run and the Vocabulary has been loaded.
}
