% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVocabMapTables.r
\name{CreateVocabMapTables}
\alias{CreateVocabMapTables}
\title{Create Vocabulary Mapping Tables.}
\usage{
CreateVocabMapTables(connectionDetails, cdmSchema, cdmVersion, sqlOnly = FALSE)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that will contain the Vocab mapping
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{The version of your CDM.  Currently "5.3" and "5.4" are supported.}

\item{sqlOnly}{A boolean that determines whether or not to perform the load or generate SQL scripts. Default is FALSE.}
}
\description{
This function creates source to source and source to standard vocab mapping tables.
}
\details{
This function assumes \cr\code{createCDMTables()} has already been run and the Vocabulary tables have been loaded.
}
