% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropSyntheaTables.r
\name{DropSyntheaTables}
\alias{DropSyntheaTables}
\title{Drop Synthea Tables.}
\usage{
DropSyntheaTables(connectionDetails, syntheaSchema, syntheaVersion)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{syntheaSchema}{The name of the database schema that contains the Synthea
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'synthea_instance.dbo'.}

\item{syntheaVersion}{The version of Synthea used to generate the csv files.  Currently "2.6.1" is supported.}
}
\description{
This function drops all Synthea tables.
}
