% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropVocabTables.r
\name{DropVocabTables}
\alias{DropVocabTables}
\title{Drop Vocabulary Tables.}
\usage{
DropVocabTables(connectionDetails, cdmSchema, cdmVersion)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the CDM schema that contains the Vocabulary.
Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{The version of your CDM.  Currently "5.3.1" and "6.0.0" are supported.}
}
\description{
This function drops Vocabulary tables in a CDM.
}
