% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadEventTables.r
\name{LoadEventTables}
\alias{LoadEventTables}
\title{Load CDM Non-Vocabulary Tables.}
\usage{
LoadEventTables (connectionDetails, cdmSchema, syntheaSchema, cdmVersion)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that will contain the different VISIT
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{syntheaSchema}{The name of the database schema that contain the Synthea
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{The version of your CDM.  Currently "5.3.1" and "6.0.0".}
}
\description{
This function loads the CDM Event tables with Synthea data.
}
\details{
This function assumes \cr\code{createCDMTables()}, \cr\code{createSyntheaTables()}, \cr\code{LoadSyntheaTables()},
             and \cr\code{LoadVocabTables()} have all been run.
}
