% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadVocabFromCsv.r
\name{LoadVocabFromCsv}
\alias{LoadVocabFromCsv}
\title{Load Vocabulary Tables From CSV Files.}
\usage{
LoadVocabFromCsv(connectionDetails, cdmSchema, vocabFileLoc)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that will contain the Vocabulary (and CDM)
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabFileLoc}{The location of the vocabulary csv files.}
}
\description{
This function populates all Vocabulary tables with data in csv files.
}
\details{
This function assumes \cr\code{createCDMTables()} has already been run.
}
