% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreCDMTables.r
\name{restoreCDMTables}
\alias{restoreCDMTables}
\title{Restore the CDM (undo the pruning).}
\usage{
restoreCDMTables(connectionDetails,cdmSchema,cdmVersion)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{The name of the database schema that contains the CDM
instance.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{cdmVersion}{The version of your CDM.  Currently "5.3.1" and "6.0.0" are supported.}
}
\description{
This function undoes the pruning by renaming (the *_orig tables are given their names back).
}
